Fix mod_ssl
Fix mod_ssl so it doesn't cause constant restarts of Apache2 service
due to duplicated definition of ports.conf with different values.

Note that apache2 1.5.0 (most recent version at this time) has a
different solution that keeps the duplicated definition but overrides
the listen_ports attribute on the node, which makes sure the *next*
run will see the same listen_ports on both definitions.

However, in that case if you ever remove mod_ssl, the node attribute
will remain set, even if you change the listen ports elsewhere. The
solution here comes from COOK-1460, with a change suggested by the
last comment (to avoid setting the ports on the node).