Increase retries and health check timeouts.
HAProxy retries only works as long as a backend is considered to be
healthy. Health checks themselves do not use "retries" configuration,
but, instead, rely on "rise" and "fall".

Given that shard, at the present, takes up to 30 seconds to restart,
I'm setting retries at 30, and configuring all other backends rise and
fall configurations to allow time enough for the service to restart.

This should only bed done if a backend does not have alternative or
backup servers, as, otherwise, it introduces a significant delay
to process. However, no secondary servers exists right now for
the services in this file.