Remove build-essential temporarily
This cookbook will be reinstalled with knife like this:

    knife cookbook site install build-essential

The advantage of doing so is the knife uses a separate branch to
track remote changes, which can be brought in while preserving
local changes by repeating the command.

The disadvantage is that we must stop updating version on metadata,
since it will be tracking the remote one.