Add ports to ports monitoring
The new recipe for apache2::ssl uses a lower priority attribute than
what was set by an older recipe, causing the override on node to be
useless. Therefore, move the override to the recipe.

Unless we reset the old node default listen_ports attribute, this
is going to be necessary for anything including apache2 with different
ports then 20000 and 443.