Fix existing haproxy munin graphs
The problem turned out to be invalid labels, something which only
the server complains about (that is, munin-run works fine). Also,
one of the scripts searched for frontends with the same name as the
backends, which is not our case.

Move frontend name to plugin configuration (it is used by another
plugin I'm testing anyway), and apply the label name substitution
recommended by Munin