Limit queries on haproxyV2
The queries are now limited to 15  for each shard, based on observed
performance. Peak is around 30, and at such times we are facing
intense memory pressure, which triggers full garbage collections which
further slow down processing leading to shard unresponsiveness.

We'll fine tune it from here. Memory improvements on shard, such as
disk-based sort and soft references for blocks, as well as changes
on the query patterns, can also lead to changes in this value.